Finally fixed problems with WITH-TIMEOUT macro
SBCL's WITH-TIMEOUT macro cannot really be nested. Even worse, deadlines are
also of type TIMEOUT. The new cram_language task backend makes heavy use of
deadlines to periodically check its message queue which then conflicts with
actionlib's WITH-TIMEOUT. The fix introduces a new macro WITH-TIMEOUT-HANDLER
that is not based on the TIMEOUT signal.


git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/roslisp_common/trunk@36201 eb33c2ac-9c88-4c90-87e0-44a10359b0c3