Not only store child-frame ids in the chache but also the frame id
In order to verify if frames are valid, we need to keep all frames (child-ids
and frame-ids) of transforms. Store the parent transform with the value 'PARENT
in the cache if not already present and overwrite it with a cache object if
the parent appears as a child frame. This allows for easily checking if frame
ids are valid by just looking them up in the cache.


git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/roslisp_common/trunk@37257 eb33c2ac-9c88-4c90-87e0-44a10359b0c3