cl-tf bugfix: the lookup algorithm would sometimes not find a valid transform:
Adding new transforms to cache entries has one corner case handler, namely, if a new transform turns out to be the first entry in the cache cell, also add it to the previous cell, as interpolation always needs a lower and upper bounds within one cell and doesn't work across different cells. However, somehow, this corner case was neglected during lookup, so if you would need a transform which would be between the first entry in a particular cell and an entry from the previous cell, lookup would simply not find it, no matter how long you would wait for transform. It would only work once you changed the timestamp. This commit fixes that.